+++ jsonpath/function_extensions/__init__.py
@@ -1,5 +1,6 @@
 # noqa: D104
 from .arguments import validate
+from .count import Count
 from .keys import keys
 from .length import length
 from .match import Match
@@ -7,6 +8,7 @@ from .search import Search
 from .value import value
 
 __all__ = (
+    "Count",
     "Match",
     "Search",
     "value",